requirements.txt
fpdf
streamlit
Pillow
supabase
